fix: Avoid pre-mature routing-keys check
The routing keys are a bit more complicated than just checking for their
existance. A few lines down from the removed check, we add to the
routing keys from the did_services_chain. Later on in the wrap function,
we check again for the existance for the routing keys before determining
whether or not we need to wrap the message so the removed lines really
are not needed.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>